Adicionando README.md + a modelagem da classe Pessoa estendendo para a classe DonoRestaurante
