+++ src/main/java/br/com/fiap/TechChallenge/model/DonoRestaurante.java
@@ -1,45 +1,12 @@
 package br.com.fiap.TechChallenge.model;
 
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
+import jakarta.persistence.Entity;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
-
-import java.util.Date;
 
 @Entity
 @Getter
 @Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class DonoRestaurante {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @NotBlank(message = "Nome é obrigatório")
-    private String nome;
-
-    @NotBlank(message = "Email é obrigatório")
-    private String email;
-
-    @NotBlank(message = "Login é obrigatório")
-    private String login;
-
-    @NotBlank(message = "Senha é obrigatório")
-    private String senha;
-
-    private String endereco;
-
-    @Temporal(TemporalType.TIMESTAMP)
-    private Date dataUltimaAlteracao;
-
-    @PrePersist
-    protected void onCreate() {
-        dataUltimaAlteracao = new Date();
-    }
+public class DonoRestaurante extends Pessoa {
 
 }
