+++ src/main/java/br/com/fiap/TechChallenge/model/Pessoa.java
@@ -1,4 +1,41 @@
 package br.com.fiap.TechChallenge.model;
 
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.Date;
+
+@Entity
+@Table(name = "pessoas")
+@Inheritance(strategy = InheritanceType.SINGLE_TABLE)
+@Getter
+@Setter
 public class Pessoa {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @NotBlank(message = "Nome é obrigatório")
+    private String nome;
+
+    @NotBlank(message = "Email é obrigatório")
+    private String email;
+
+    @NotBlank(message = "Login é obrigatório")
+    private String login;
+
+    @NotBlank(message = "Senha é obrigatório")
+    private String senha;
+
+    private String endereco;
+
+    @Temporal(TemporalType.TIMESTAMP)
+    private Date dataUltimaAlteracao;
+
+    @PrePersist
+    protected void onCreate() {
+        dataUltimaAlteracao = new Date();
+    }
 }
